Added global variable to change the focusing behaviour of HWindow-atteched views and controls.
When the focus behaviour is 1, clicking on any subview brings the window to front, if attached to a HWindow instance.

If the behaviour is 0, only direct clicks on the HWindow controls brings the window to front.

The default behaviour is now 1, the previous defaulted to the same behaviour as 0 would do now.




git-svn-id: http://svn.rsence.org@395 5bb182ed-6ace-4cc9-85f1-9567b96542ab